Generate multiple StatsJob.
Instead of enqueue only one statsjob every X seconds, we enqueue N
jobs, where N is the number of different services that request to our
api at an interval of X seconds.

With this way, we try to distribute the load of generate stats and we
try to take advantage of have a lot of available workers.