Update code to work with a different redis for queues.
Our next big step will be the redis sharding. For it, I think that have
the queues in the shard is not a good idea. With queues in the shard,
one node always receives more requests.

I added a dependency with redis-sentinel gem to manage the redis
failover. I updated the configuration and the deploy files to work with
sentinel properly.