memoizer: boost performance when invoking from metaclasses
This commit changes the way in which class methods
are accessed so that we always get a bound reference
to them from within an instance context. This way we
fix the performance regression when memoize is used
inside a metaclass context to memoize class methods
(ie. instance methods in a metaclass).