application: stop inheriting from Core
This commit introduces all the features found in Core so that the
Application class stops depending on that gem. Since the thing is
fscking complicated when imported all at once, I've tried to
carefully divide functionality between what already was there and
what we are bringing in with Core: this way it is easier to review
and understand.

Once the damn thing is considered stable I will work on merging all
modules into a single class.