internal_api/services: fix possible method call on nil value
Service.load_by_id may return a nil if it does not find the referred
service. We don't check the return value and directly try to use it
as a service object. Fix this by using the bang method which raises
in case the service id does not exist and in that case return a 404.