Update Cache#clean_cached_xml method.
Only check limit violations on metrics with changed values.

This solves the current issue, but it is not the best approach. We
need to check the possible impact that this could have in production
performance.
With this way, if some metric is out of quota, we are not taking
advantage of pregenerated xml because we are ignoring it and
validating the status every time.
A possible better approach could be store a pregenerated xml
depending on the requested usage. But this solution would suppose
more code and more testing. It will be our plan B.