internal_api: introduce single Sinatra app support
The new internal_api method in ThreeScale::Backend::API receives
a route and a block in which to define typical Sinatra HTTP methods.

The effect of this is that the route specified is included in the
single Sinatra application architecture we have been discussing in
backend, _AND_ that any methods you define will end up in that
application, so that:

* You should try to avoid defining behaviour inside the block.

In particular, methods you define in there are going to be declared
at class level inside the Internal API Sinatra app. And you don't
really want those methods clashing with other methods that other
parts of the API define. You can avoid that by using static methods
from external classes. If you need state you can as well embed it in
external objects.