metric: make .delete return a boolean
This returns false when one of the required keys is not found. Of
course, we are not exhaustively testing this, because, well, no
friggin code path is doing that. Instead this is meant as quick
check to know whether a given metric exists before being deleted or
not and then deletion is meaningless.

The only real user is the internal API.