internal/usagelimits: introduce UsageLimits internal API
This just implements the currently needed methods by Multitenant.

This is **FAR** from being a complete microservice at this point.
For a nice API we should also include other endpoints, including
those to get collections. Unfortunately that work is currently not
easily justifiable, since we would not have users at this point.