app/api/internal: catch accesses to non-existent routes
Previously Sinatra returned 404 with a JSON content-type but an HTML
message due to the way it was set up. Now we catch 404's provoked by
accessing routes that Sinatra does know nothing about by way of
intercepting Sinatra::NotFound, and return actual JSON.

This fixes an error when "searching" for an empty stringed end user
in multinenant which was accessing an unknown route.