Dockerfile: use new quay ruby image since it's what everyone will use
It's supposed to be lighter, more updated and more maintained.

Note: you NEED to authenticate to quay.io. There's no way around
that, so either you provide your credentials when Docker runs,
interactively, or you create a .dockercfg file in your $HOME that
quay.io will gladly generate for you in the repo, with a KEY to read
from the repo. To do that, you go to the repo in quay, go to
settings, and in the dropdown box select generate a robot account,
name it your username, and download the config... And "bam, done!".