application: DRY up params checking for app_id, user_key, access_token
Here we had an inconsistent behaviour:

- If we got a valid user_key, but the application was not found,
  we reported an invalid user_key.
- If we got a valid access token, but the application wasn't found,
  we reported application not found.

IMO the last action is the one that makes sense: you should not
report an invalid user key when it is valid, but the app it points
to is no longer existing. This commit changes to that behaviour, and
thus changes the tests involved so that we expect the right
exception in those conditions.