Update some tests in LatestEventsTest
Enjoy: These tests were checking that we do a request to events_hook
when processing a report transaction, changing the events_hook to a
invalid URL and assuming:
 - net http request will fail
 - we're rescuing exceptions in that point and sending and Airbrake with
 exception info
 - airbrake is not configured at that point and it's generating an
 exception: NoMethodError: undefined method `send_to_airbrake' for
 nil:NilClass

Awesome.

This commit fixes these tests. If we want to check that after a report
transaction, a request to events_hook is triggered, we only need to make
an expectation to check that we call to `request_to_events_hook`.