usage_limit: boost performance in hot path method by 45%
It wasn't clear to anyone that this was just a product matrix very
much manually implemented. Ruby already has Array#product which
benefits from a native implementation and less VM churn.

This change allows us to get rid of the `pairs_of_metric_id_and_period`
method. Find the benchmark below.

These are the results for different implementations of the
`pairs_of_metric_id_and_period` method. 'original' is the code before
this commit, 'reduce' is a cleaner version using inject/reduce/fold,
and 'product' is just using Array#product.

Calculating -------------------------------------
            original     3.992k i/100ms
              reduce     3.815k i/100ms
             product     5.676k i/100ms
-------------------------------------------------
            original     40.996k (± 0.5%) i/s -    207.584k
              reduce     38.846k (± 0.7%) i/s -    194.565k
             product     59.589k (± 0.5%) i/s -    300.828k

That's a ~45% boost for this hot path method, plus getting rid of it
altogether.