usage_limit: boost performance of partial hot path by 50%+ (worst case)
The rest of the `load_all` method maps keys and values and creates
new objects. The boost relates to the worst case, which is when we
get no nil values.

This is the benchmark in the worst case:

Calculating -------------------------------------
            original     1.875k i/100ms
                 new     2.766k i/100ms
-------------------------------------------------
            original     19.013k (± 1.1%) i/s -     95.625k
                 new     28.258k (± 1.1%) i/s -    143.832k

When we get nil values, the performance of the new code increases
exponentially with the amount of nils, while the original performs
linearly. This is the same benchmark with ~90% nils:

Calculating -------------------------------------
            original     3.410k i/100ms
                 new     9.039k i/100ms
-------------------------------------------------
            original     35.466k (± 0.4%) i/s -    177.320k
                 new     96.476k (± 0.4%) i/s -    488.106k

IOW, the original code complexity was O(n) *wrt nils*, and this is
O(log n).