listener: fix Swagger misapplied code tag breaking the render
The <code> tag was applied to a tag instead of an HTTP code, and it
rendered parts of the doc with red color. Since the render did not
find the matching closing </code>, it also rendered broken
descriptions of parameters (too much to the right).

Fixed by just not using <code> in the wrong places.