server: special-case thin instead of using Rack for it
This is because running Thin through the Rack code regresses a lot.
In particular, Rack has its own view about daemonization and
logging, and adapting a webserver which already provides this out of
the box to Rack means some extra work that's just not worth it.

The thing is that Rack should be used for NON-PRODUCTION servers
testing. It is not designed to handle by itself the complexity of
running and managing servers in production.