transactor: catch JSON::GeneratorError when enqueuing invalid UTF-8 for logs
Unfortunately, that crappy software named Resque was silently using
Yajl for encoding JSON. It would swallow malformed UTF-8 and explode
on parsing, for which we added some code to just leave an error
message in the log record.

This commit moves that handling to the enqueuing time, since JSON
will barf on encoding time as opposed to Yajl.