Add more specs for AlertLimit.delete
We're not checking the value when we're deleting an alert limit, and if
you send nil or "fooo" as a value, when we try to delete it we are
casting it to integer. nil.to_i or "foo".to_i == 0, so we would be
removing the alert limit with value 0 and that isn't the valid
behaviour.

With these specs, we're testing these cases.