Load service_id when we store errors
Sometimes, an error is raised before we load @service_id so we can't use
it. We need to call to service_id method to load it explicitly. Then,
some errors as wrong service_id or invalid provider key can happen and
we don't want to raise them (we want to raise the original one), so we need
to capture and ignore them.