gemspec: make it possible to deploy using bundler from gem install
This patch creates a gem extension to make it possible to install
and run `Bundler` on the `Gemfile` at gem installation time.

This is so because the current deploy method involves creating
a gem, uploading it to a geminabox server and having servers fetch
the software from there through `gem install`. Unfortunately this
method totally ignores Bundler's `Gemfile`, which is a much more
powerful and flexible way to deploy software.

The extension is just plain Ruby code to make sure Bundler is
installed and use it to install (system-wide) the dependencies
specified there.

The upside of this is that we can transition now to this method of
dependency management without having to wait for the project and the
tooling to support it directly.