server/puma: set STD{OUT,ERR} redirects to log file if specified
This sets the redirects to the log file specified (if any). Note
that this has to monkey patch Puma::CLI#parse_options because there
is no way to access the settings after they are loaded and before
they are applied. This probably warrants a PR for puma/puma, but
meanwhile we just make sure this works by checking the Puma version.