spec/internal: replace app instance variable with lets for readability
I noticed that some internal api tests (application_keys,
application_referrer_filters, application) define the variable @app
inside a 'before' block instead of using a let(:app) or a
let!(:app). I realized that this is because this line: set_app
ThreeScale::Backend::API::Internal sets the variable app and it's
needed to make the tests work. let(:app) or let!(:app) overwrite app
and that makes the tests fail. This is probably why it was decided
to use an instance variable(@app).

I think this is too confusing and we should use 'let' with a
different name.