config.ru: use require_relative bc at this point we have no proper $LOAD_PATH
By the time this gets loaded in Puma, the process is a fork from the
Puma master server process. The idea there is that we have pruned
all Bundler environment and only have Puma and its dependencies in
the load path, so we will not have a '3scale_backend' load path.