logger/middleware: default request logger to STDOUT
In some cases the logging middleware would not receive a logger
parameter and use the 'rack.errors' stream instead for logging
requests, which is not really ok for our middleware (it would be
nice though to use that stream for outputting backtraces).

We now default to STDOUT unless another value was given. If that
value was nil, we resort to 'rack.errors' and if that is not
available we fall back to STDERR.