listener: support haproxy endpoint
This commit adds support for the `HEAD /haproxy` endpoint. This is
used by upstream HAProxies to check availability of other HAProxies
but we can use it also ourselves to assert that we are alive.

Currently we just return a 200, but this opens the door to on-demand
availability switching, thus effectively removing (or adding)
servers to upstream proxies pools.

My idea for the near future is to allow this endpoint to be changed
dynamically by some unspecified means and have the init.d script
call it to switch off availability when stopping, including when
fully restarting an application server.