errors: fix ServiceRequiresDefaultUserPlan airbrake
This was being called with an id. Which should probably be what has
to be done, but that would imply that either we have no use for the
parameter or we do and change the error message reported back to
users thus incurring into a minor form of breaking change.

So since I don't have a strong opinion about this either way I have
just opted for fixing this and let the message stay the same. This
should however be reviewed for a future API version.