Add stats parser
We are using keys like this one to store stats in Redis:
"stats/{service:1006}/cinstance:a60/metric:255/week:20151130"
This parser accepts a key like that and its value, and
constructs a Hash. That Hash is going to be what we are going
to send to the Kinesis adapter that we will develop later.