Add stats bucket manager
This class will help to simplify other classes that we need to
implement the background job that will send events to Kinesis.
For example, while I was developing the BucketReader class,
I noticed that I needed to perform many calls to storage.X.
Now I will be able to use this class and simplify code.