BucketReader: return the latest bucket read in #pending_events_in_buckets
Without knowing the latest bucket read, the caller of #pending_events_in_buckets
cannot set the latest_bucket_read. The caller is the one that needs to be
responsible for this, because we cannot take for granted that once a bucket has
been read, all its events will be processed correctly.