KinesisAdapter: implement basic batching strategy
To simplify thing for the proof of concept, we are going to use a
basic batching strategy. This strategy is based on the number of
events of each record and batch instead of the size in bytes.
We will implement more complex batching strategies later.
In a later commit, we will deal with the different types of errors
that Kinesis can return.