SendToKinesisJob: delete buckets as they are read
We need to delete the buckets at some point after we make sure that
they have been read. There are several strategies that we could follow,
the easiest one is to make the job responsible for deleting them
after it has read the events of the buckets and sent them to Kinesis.
That is the strategy implemented in this commit, but it can change
in the future. For example, we might decide to have a job that deletes
buckets periodically.