KinesisAdapter: send data to Kinesis in a format that can be read by Redshift
Instead of sending a correct JSON, we are going to send a pseudo json.
This contains the events in correct JSON but without using '[]' or
separating them by commas. This is one of the formats that Redshift can
read easily. See: http://docs.aws.amazon.com/redshift/latest/dg/copy-usage_notes-copy-from-json.html