Makefile: revamp to support Docker persistent containers and versioned images
This is kind of a huge change to how we use our Makefile right now.

With this change, we can easily support multiple Ruby versions. The
motivation is to ease the deploy of backend with Docker. After landing
this we can have $RUBY_VERSION$ placeholder in Dockerfiles that can be
used to select specific images or install Ruby packages. The idea is
that upgrading the Ruby interpreter would ideally be a matter of
changing the contents of .ruby-version.

This also adds support for versioned Docker images, ie. backend:2.2.3 vs
backend:2.1.6. This is useful for quite a few things, such as testing
different versions in Jenkins and in parallel.

And finally, there is now a `make dev` target that creates a persistent
container for development, so that our changes are not lost.