BucketStorage: increment number of keys for each mget
I tried several values for this constant in production. For the amount of
keys generated in around 2min 30s, the runtime of the method
buckets_content_with_values was reduced from around ~1.5s to ~0.6s.

We could increment it some more, with KEYS_SLICE_CALL_TO_REDIS = 3000, the
runtime was ~0.3s, but it could have an impact in other areas of the code.
We would be blocking Redis for more time (cannot perform other ops while
running an mget).