SendToKinesisJob: filter events before parsing them. Improves perf
For the amount of keys generated in around 2min 30s in production, the runtime
of the method prepare_events went from ~4.7s to ~3.4s.
Filtering the events now deals with strings instead of hashes and it is a bit
slower because it has to run .include? for each of the keys. However,
it pays off because parsing the events is by far the most costly operation
inside the prepare method and now it has to parse less keys.