Makefile: make sure test always goes through the Dockerfile
Without this, changing dependencies both at the container and gem
levels would not work.

This also changes the name of a function so that it is clear that we
just run a container, since when testing the container has already
processed the Dockerfile and it already contains the volumes. This
way the original checkout in test does not get modified, which is
a nice thing.