KinesisAdapter: rescue kinesis client errors
We can receive different types of errors when sending to Kinesis. For example,
if we exceed the amount of data that we can send for a given period of time,
we will get a Aws::Firehose::Errors::LimitExceededException.
We want to rescue those exceptions and save the failed batches of events
for later.