ResponseCode: send response codes stats to the correct buckets
Response codes are treated in a different place than the rest of
metrics. However, they are stored in the same way in Redis, using
the same kind of 'stats' key.

The problem is that we were sending the usage metrics to the
appropriate buckets but not the response codes.
The name of the bucket is determined by Keys.changed_keys_bucket_key,
but we were not calling that function.