Worker: do not die when decoding invalid arguments for job
This is a quick-fix to avoid a disaster. It is not the ideal
solution. The problem is that after getting a job from the resque
queue, the worker decodes it and if the decoding fails, the worker
process dies.

The decode method can raise a Resque::Helpers::DecodeException, but
we are not 'rescueing' from it.

Of course, this is not acceptable. The job should not fail.

As this is a quick-fix, for now, we are just going to ignore those
jobs. It is better than ignoring them and crashing. However, we
should log the problematic jobs somewhere, or save them in a
different queue to be processed later.