Worker: do not discard jobs whose args contain no UTF-8 chars
We were discarding all these jobs before. I think a better option
is to pass the strings with invalid UTF-8 chars to the jobs and
let them decide what to do.

For example, the job responsible for storing logs, saves an alert
that indicates to the user that he sent a string that contains non-UTF8
chars.