metric: remove reverse mapping left-overs when renaming a metric
Simply by renaming a metric (ie. saving metric data for an existing
metric id with a different name) makes both the old and the new
names point to the same id in backend, which allows you to refer to
the old name in calls to BE although it does not exist anymore,
leaking also the mapping in Redis.