oauth/token_storage: reimplement token storage management
This is the real meat.

We change the way OAuth tokens are stored so that now we can
associate both an application id and a user id. The latter is
optional, so we have to take that into account.

The way that these are now treated is using lazy enumerators so that
this code should be ready to work on partial lists of tokens (which
may come later as we substitute SMEMBERS by SSCAN in Redis) and stop
being a memory hog when huge quantities of tokens come in.

The really awful bits deal with arrays of:

1. [token, key, value, ttl]

and

2. [[tokens], [keys], [values], [ttls]]

Creating _just_ `Token` objects would not be good for performance
because we need to support both kinds of arrays:

Array kind 1 is ideal for things like filtering tokens, but kind
2 is ideal for other reasons, such as pipelining accesses to Redis.

There are in fact other kinds of arrays handled internally dealing
with the lazy enums, but those are not important here.

This code basically needs the tests to be changed to adapt to the
new expectations about the storage providing the user ids as well as
application ids.