NotifyBatcher: delete misleading comment
This comment no longer applies. It makes no sense to modify
configuration.notification_batch directly in the code just so the tests pass.

This config variable should be set in the tests where it applies, like we are
currently doing in /test/test_helper.rb and /test/unit/transactor/notify_batcher_test.rb,
for example.