aggregator: increase limit of buckets that can be created
We have checked in production that the current limit is too low.
The memory needed by 60 buckets is too low to cause our Redis any issues.
I am increasing the limit to 360 buckets, which is an hour of data.
This number should not cause any issues, and also, it gives the
system more time to recover if there are issues.