bucket reader: take into account only buckets that exist in storage
We now allow specifying a max number of buckets to retrieve the events from
instead of getting the events of all the existing buckets.

To support this, we need to change the way we calculate the pending buckets.
Before, we had a method that calculated the pending buckets without accessing
Redis. It simply grabbed a start and end dates, and calculated all the
buckets generated in the interval, without checking if they exists in Redis.
This works OK when we want to retrieve all the pending buckets. The problem
is that if we have not read buckets for some time, and we asked just for a
few of them. Imagine that, for some reason, we have not read any buckets in
a day, and the only ones that exist in Redis are the ones from the last few
minutes (we might be playing in preview disabling/enabling the bucket storage).
If we only asked for 60 buckets (reasonable, 10 mins) in a job, we would
need to execute the job many times until we retrieve the existing buckets, and
most times we would not be getting any real buckets. This makes no sense.
This is why now we only return buckets that exist in Redis.