transactor: raise ServiceIdInvalid if service exists in other provider
This fixes a bug. Before, in the authorize methods, when we received a valid
provider key and a valid service ID but belonging to a different provider,
we raised InvalidProviderKey. We should raise InvalidServiceId. Otherwise,
we might confuse users. They are being told that their provider key is invalid,
when it is not.