application events spec: fix first daily traffic test
The problem is that the test only works if Time.now gives the same result in
let(:event) and inside the first_daily_traffic method.
We need to use Timecop.freeze to force that. Otherwise, this might fail if for
some reason the test is not executed fast enough. Apparently, this has happened
to some people.