redshift adapter: fix import of events
There was a bug. Some events were imported even when they should not be. The
reason is that we made sure to delete all the existing events that had a newer
version in the table of events that we needed to import. However, when there
was an older version in the table of events that we needed to import, we
imported it anyways, thus inserting duplicated events. This commit fixes that
issue.