FailedJobsScheduler: ensure that failed jobs are rescheduled only once
We need to use a distributed lock. The reason is that the rake task that
reschedules failed jobs is executed periodically from several worker servers
and there is the risk that some of them will try to execute it at the same
time. This could result in rescheduling the same failed job several times.

To understand this, remember that the call to Resque::Failure.requeue does not
delete a failed job from the queue of failed jobs.