redshift adapter: optimize query by using smaller table in join
In the query that deletes from the imported events table the ones that do not
need to be imported because there is already a more updated version in the DB,
we were performing a join with the whole 'events' table. This query can become
too slow when the 'events' table grows. We can achieve the same thing just by
selecting a few rows of the 'events' table based on their 'time_gen'.