Rakefile: rescue from DecodeException when rescheduling failed jobs
The rake task that reschedules failed jobs crashes when there is a job with
an invalid utf-8 encoding. This is a problem because it can re-enqueue some
jobs until it finds the one with invalid encoding, and then crash before
clearing from the failed jobs queue the ones that have been re-enqueued.
This means that those jobs will be rescheduled again the next time the
task is executed.

This is a quick-fix that simply discards the jobs with a non-utf8 encoding.
A more sophisticated fix will be introduced in the future.