redshift adapter spec: add integration tests using dev Postgres
This commit replaces the old unit tests with integration tests. The new tests
use a Postgres that runs in our dev Docker container. The previous tests were
OK at first when the logic and the SQL used were pretty simple. However, some
SQL statements got more complex after optimizing some things. At that point,
it became evident that we needed to use integration tests to be confident that
our code works as it should.

We use Postgres because Redshift is based on it. There are some differences
between them, for example Postgres does not support the COPY command. We take
those things into account in this commit.