resque hacks: avoid allocating unnecessary objects.
Resque calls the hook methods (before_hooks, around_hooks, etc.) each time a
resque job is executed. In each of those methods we were creating unnecessary
objects because they perform greps and sorts and the result never changes.

This commit caches those results. It only calculates them once per execution.
This makes a big difference in the number of objects that we allocate in
worker processes. We measured this using Stackprof and around ~10-15% of
the object allocations we were making were done in those hook methods.