validators/referrer: move validator logic from Transactor
There was some logic that belongs to the referrer validator inside the
Transactor. This commit moves that logic where it should have been implemented.

The unit tests of the referrer validator did not take into account the logic
that was in the Transactor. This is why the tests have been modified. In
practice, this means putting the application used in the tests in a service
that requires referrer filters.