Remove non-representative benchmarks
These benchmarks are not representative of what we are running in production.
There are several reasons for that: the benchmarks use a local Redis instance,
they are not executed under a heavy load, they only use a service with a few
metrics and usage limits, etc. For all those reasons, the results obtained by
running these benchmarks can be misleading, and it is dangerous to draw
conclusions from them.

We never run these benchmarks currently, anyway. Also, we have better ways of
measuring performance now: github's scientist gem, rbtrace, etc.