integration tests: make test class names unique
We had several tests in different files defined using the same class name.
The problem with this is that the 'setup' method is redefined so the behavior
of these tests ends up being really counter-intuitive. Luckily for us,
the tests with same name seem to have a very similar or identical 'setup'
method so the tests passed anyway...