errors test helper: do not call internal API endpoints
The methods in the errors test helper were calling private API endpoints. These
will disappear when MT update to a newer version of the core gem.
There was another problem with #assert_not_errors_in_transactions: it only
verified that there were no errors for the default service associated with a
provider key instead of all of them. We need to pass a list of service IDs
instead of a provider key.