cache: make .caching_enabled? return false always
The Cache module was introduced several years ago, and while it
could have been effective back then, we have no evidence that it is
now given the current codebase and workload.

We decided to evaluate its effectiveness because it is probably one
of the most complex parts of our codebase. When we need to modify
something in the Transactor module, we always need to keep in mind
that the cache might affect the results. This is slowing us down.

We found out that, apart from being hard to maintain and reason
about, the Cache module has a negative impact in both listener
latencies and background job runtimes. With the cache disabled,
listener latencies are 15-20% lower for percentiles 90 and 95. The
improvement is a bit better for average latencies. Regarding the
background jobs, their runtime is reduced by a 20-25%.