transactor{,/status}: parameter "hierarchy" emits info on metric hierarchy
This adds the "hierarchy" tag to an authorization response informing
the client that the limits specified in the usage reports also apply
to the set of metrics for which there is a parent-children relation,
which means that for usage limit purposes the limits specified MAY
further restrict those directly applying to those metrics (even if
those metrics have none), and that extra computation needs to take
place to determine the real limits.

This has been implemented to give users a way to identify which
metrics not being listed in usage reports also have indirect usage
limits.