aggregator: avoid unnecessary call to limits validator
The limits validator only mutates the status when limits are exceeded. It does
that by calling Status#reject!. That only modifies these attrs: authorized,
rejection_reason_code, and rejection_reason_text. Those are not used in the rest
of the aggregator code, so it's safe to remove the call to the validator.
This is important because the aggregator is called in every report job.