listener: don't return bare error codes in OAuth mgmt endpoints
This should make it so that error conditions in such endpoints don't
just return an error status, but also the corresponding error data
in the body (because we are now using exceptions).

While at it also stop accessing twice `params` for some data that is
accessed several times such as `service_id` or `app_id`.