listener: rename extensions methods to avoid overriding Sinatra's methods
We need to rename Listener.extensions. Sinatra already defines that
method and we are overriding it. I'm not really sure about all the
consequences of that. The most obvious one is that we are calling
register AllowMethods and if we do not override the method,
Listener.extensions returns [ThreeScale::Backend::AllowMethods].