Rakefile: show swagger tasks only in SaaS
These tasks are used to generate the swagger docs that document the backend endpoints and upload them to the 3scale support website.
This should not be included in the on-premises version as the only one that needs to update the endpoints documentation in the support website is the 3scale team.