backend: force event buckets creation to false when not in SaaS
Buckets are only used in our SaaS analytics system. We need to force
the option to false when in on-prem. This way we will not need to
explicitly set the value for can_create_buckets when on-prem and
will have the correct value even if we set an invalid one by
accident.

We will come for a better solution to handle this kind of
dependencies in the future, but this is good enough for now.