backend: change substitute file for Airbrake.notify when not configured
This way we can use STDOUT as a substitute when we are on-premises
without Airbrake. Also, we can write to /dev/null while running the
test suite. Ideally, we'd like to set the logger from the
test_helper and the spec_helper files. However, that is not possible
because the logger is initialized when 3scale/backend is required.