spec/unit/service: add tests for .load_with_provider_key!
This method was a private method of Transactor, so it didn't have any unit tests.

I think it's necessary to have these tests to have the errors returned in every
documented case. We might want to change some of them. Also, the method looks
rather complex, so we might want to simplify it. This commit it's just a first
step: writing specs for what we already have.