spec/unit/service: bring tests from test/unit/service
We had tests for the Service class both in the test-unit suite and Rspec. This makes little sense and there were only a few tests in the test-unit suite, so I decided to migrate them to the rspec one.
Note that I did not adapt the test that checks that the load method is memoized. That is testing a detail of the implementation and not functionality. Those kind of tests belong to the unit tests of the Memoizer.