spec/unit/failed_jobs_scheduler: fix tests that call requeue()
These tests were not working correctly. When ".to receive()" is used and "and_call_original()" is not called, the method is not called. In the case of requeue() is necessary because sometimes exceptions are raised. When the original method is not called, those exceptions are not raised and the tests do not check what they are supposed to.