Gemfile: point Puma to the 3scale-patched version
When serving non multithread safe applications, a single worker thread
is used, possibly with quite a few worker processes.

When a keep-alive request is dispatched, Puma checks for additional data
(a new request) coming in. This checks makes any one worker thread stall
if no data is available, up to FAST_TRACK_KA_TIMEOUT seconds (set to
200ms currently).

This has the effect of neglecting other requests that are ready, waiting
to be picked up for processing.

This patch fixes the case for max_threads == 1. However, my intuition
suggests we should limit this feature considerably to avoid starvation
and denial of service, so I expect further patches to avoid having all
available threads blocked on KA clients.

Even having a fast check with a timeout of 0ms opens the possibility of
someone sending one KA connection per thread and starving other
legitimate connections in the pool. I think capping the number of
threads available to serve KA connections even when they are fast enough
is a desirable property.