ext/mkrf_conf: use the Bundler version that created the lockfile
Currently the gem install process would install a fixed version of
Bundler, 1.11.2. This was ok before, on SaaS-only and because of
certain problems with Bundler, but now that we have the BUNDLED WITH
information in the lock file we can use that to ensure the same
Bundler version is going to be installed and used.

This is only applicable to direct installation with gem install, so
not related to on-prem because we use bundler directly to install it
(and we don't plan to use the gem for installing it).