openshift/Makefile: build in a separate directory
This introduces a few targets so that the build process for the
openshift image is done in a clean directory. The advantage is that we
don't need to keep populating a dockerignore file for this specific task
and that we will be able to stop using gem in the Dockerfile and use
Bundler for everything.