Rakefile: check SaaS before adding tasks from a testable env
This is so that we do not try to load RSpec or add tasks that cannot be
done without the test, spec or docs directories. Those might come in
later and if so we can make checks for RSpec and/or the directories, but
as of now we do not ship that yet.

This will likely need a fair rework so that we clearly define what
should be included and what should not _and_ we properly detect each
case with the corresponding features.