Extract Writer class from Middleware
The new Writer class contains the code that will be shared by all the log writers (json, apache-based). It's extracted directly from the Middleware class. I just split some methods to help readability.
This commit also changes the interface of the Middleware class so it accepts an array of writers instead of a logger.