period: do not include Enumerable
Some external code thinks of Period instances as collections that
can be cheaply enumerated, which is not always the case, especially
when this code enumerates deeply, ie. think about breaking down
:eternity until you have a collection of all the seconds.

This is what I have found with RSpec when it fails an expectation
and wants to print it out.

Since we don't really need this, we just provide :each as an alias
of break_down (which you can use anyway to get a collection that
really implements Enumerable so you can still call other typical
methods).