usage_report: use hierarchy to make sure children add to parent when rendering
We identified and fixed the problem described below:

Suppose that we have a metric M1 and a metric M2, and M2 is a child
of M1. That means that calling the report endpoint with a usage of
M2: 1 will increase the usage of M2 by 1, but will also increase the
usage of M1 by 1.

Suppose that we have defined any kind of limit for those 2 metrics.
That means that they will be shown in the response of the authorize
and authrep endpoints. The response XML contains, for each limited
period, the current usage and the maximum usage allowed, among other
things.

Now, imagine that the current usage of M1 and M2 is 0, and we
perform an authrep with a usage of M2: 1. What should be the current
usage shown in the response XML? Intuitively, it should be 1 both
for M1 and M2, but one could also argue that it should be 0, because
reports are asynchronous so they might not be still taken into
account when the user receives the response. I don't think this is
documented anywhere. I guess both approaches would be valid if
justified, documented, and agreed upon.

The problem is that what we are currently returning in the response
is inconsistent. We are returning a current usage of 1 for M2 and 0
for M1. The fact that increasing a child metric also increases the
usage of the parent is not being taken into account. Notice that the
problem only affects the response of authrep endpoints.